test: refactor test_cache.py to reduce boilerplate and decouple app tests
- Moved `TestCacheDataFiltering` to `tests/test_app.py` to separate app layer tests from cache storage tests.
- Extracted a `populated_cache_manager` fixture to eliminate 20+ instances of duplicate data setup boilerplate.
- Extracted an `age_cache_tier` helper function to cleanly manipulate cache JSON metadata and artificially age cache tiers.
- Replaced a flaky `time.sleep()` call in `test_partial_refresh_updates_metadata` with explicit timestamp manipulation.
- Hoisted inline imports in test methods to the file level to adhere to module conventions.
